requirements.txt
flask
colorlog
requests
psycopg2-binary
cryptography
gunicorn
python-dotenv
schedule
